+++ CHANGELOG.md
@@ -9,6 +9,11 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Added
+
+- Added new field `Job.result`, `Job.execution.result` is deprecated.
+- Added new field `Task.replica_rank`.
+
 ## [v1.17.4](https://github.com/allenai/beaker-py/releases/tag/v1.17.4) - 2023-02-17
 
 ### Fixed